need to understand why is dividing the first line on the map in two WTF

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: dcella-d <dcella-d@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 12:14:30 by dcella-d          #+#    #+#             */
-/*   Updated: 2023/05/31 19:07:13 by dcella-d         ###   ########.fr       */
+/*   Updated: 2023/06/01 17:28:33 by dcella-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ int		ft_isalpha(int c);
 
 int		ft_isascii(int c);
 
-size_t	ft_strlen(const char *str);
+int		ft_strlen(const char *str);
 
 int		ft_isdigit(int c);
